added field index, updated send plant data
Field now records their row and column indexes when they are generated, but somehow the numbers are a bit off, need to look into this a bit more.
Sending plant data now finds the type of the plant as tag of the object, and the row and column index of the field, and sends them to database. Need to update status based on moisture, temperature, and pressure.